+++ shared/dev-web-links.js
@@ -13,24 +13,29 @@ Object.defineProperty(exports, "__esModule", { value: true });
  */
 var DevWebLinks;
 (function (DevWebLinks) {
-    var IS_LOCAL = false;
-    var LOCAL_PORT = 9000;
-    var SERVER = !IS_LOCAL
-        ? 'https://github.com/a-bentofreire/__REPO__/blog/master/'
-        : "http://localhost:" + LOCAL_PORT + "/__REPO__/";
-    var RAW_SERVER = !IS_LOCAL
-        ? 'https://raw.githubusercontent.com/a-bentofreire/__REPO__/master/'
-        : "http://localhost:" + LOCAL_PORT + "/__REPO__/";
-    var GALLERY_RELEASE_SERVER = !IS_LOCAL
-        ? 'https://a-bentofreire.github.io/abeamer-gallery-release'
-        : "http://localhost:" + LOCAL_PORT + "/__REPO__/";
-    var MAIN = !IS_LOCAL ? 'abeamer' : '';
-    var MAIN_PREFIX = !IS_LOCAL ? 'abeamer-' : '';
-    DevWebLinks.MAIN_REPO = SERVER.replace(/__REPO__/, MAIN).replace(/\/\/$/, '/');
-    DevWebLinks.MAIN_REPO_RAW = RAW_SERVER.replace(/__REPO__/, MAIN).replace(/\/\/$/, '/');
-    DevWebLinks.RELEASE_REPO = SERVER.replace(/__REPO__/, MAIN_PREFIX + "release");
-    DevWebLinks.RELEASE_REPO_RAW = RAW_SERVER.replace(/__REPO__/, MAIN_PREFIX + "release");
-    DevWebLinks.GALLERY_RELEASE_REPO = GALLERY_RELEASE_SERVER + '/';
-    DevWebLinks.GALLERY_RELEASE_REPO_RAW = DevWebLinks.GALLERY_RELEASE_REPO;
+    /**
+     * Initializes the web links based if isLocal or not.
+     */
+    function setup(isLocal) {
+        var LOCAL_PORT = 9000;
+        var server = !isLocal
+            ? 'https://github.com/a-bentofreire/__REPO__/blog/master/'
+            : "http://localhost:" + LOCAL_PORT + "/__REPO__/";
+        var rawServer = !isLocal
+            ? 'https://raw.githubusercontent.com/a-bentofreire/__REPO__/master/'
+            : "http://localhost:" + LOCAL_PORT + "/__REPO__/";
+        var galleryReleaseServer = !isLocal
+            ? 'https://a-bentofreire.github.io/abeamer-gallery-release'
+            : "http://localhost:" + LOCAL_PORT + "/gallery-release";
+        var main = !isLocal ? 'abeamer' : '';
+        var mainPrefix = !isLocal ? 'abeamer-' : '';
+        DevWebLinks.mainRepo = server.replace(/__REPO__/, main).replace(/\/\/$/, '/');
+        DevWebLinks.mainRepoRaw = rawServer.replace(/__REPO__/, main).replace(/\/\/$/, '/');
+        DevWebLinks.releaseRepo = server.replace(/__REPO__/, mainPrefix + "release");
+        DevWebLinks.releaseRepoRaw = rawServer.replace(/__REPO__/, mainPrefix + "release");
+        DevWebLinks.galleryReleaseRepo = galleryReleaseServer + '/';
+        DevWebLinks.galleryReleaseRepoRaw = DevWebLinks.galleryReleaseRepo;
+    }
+    DevWebLinks.setup = setup;
 })(DevWebLinks = exports.DevWebLinks || (exports.DevWebLinks = {}));
 //# sourceMappingURL=dev-web-links.js.map