+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - abeamer serve --list-dir also prints the link with `http://localhost:${port}/?dir`
 - abeamer serve --list-dir sorts the listing, and places directory names first.
 - Class property animation.
+- gulp build-docs* and build-gallery-release with local and online links.
 ### Fixed
 - Fixes gallery/localization text label.